requirements.txt
fastapi
uvicorn[standard]
apache-libcloud
python-dotenv
pydantic-settings
pydantic
python-multipart